requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
rich>=13.7.0
tqdm>=4.66.1
requests>=2.31.0 